pkg/ansible/proxy: refactor proxy to remove public exposed functions (#1413)
